requirements.txt
PyJWT==2.6.0
coverage==7.1.0
pytest==7.2.1
pytest-mock==3.10.0
requests==2.28.2
